rename: move Input component to common folder
